Added the flip action for the map view.
For a tutorial:
http://stackoverflow.com/questions/843534/flip-view-iphone/2106115#2106115

Added a Nib file for the map view.
Within the nib, created a Map Controller.
Wrote the Map Controller
Did the appropriate links
Created a method and a selector for the button at the top of the view for the flip action
Followed steps in above link to make the flip happen.

Still need to make the flip come back if possible.

Reworking the data model just for kicks