May have forgotten the Map View nibs etc.
Added a DishDetailViewController and Nib files. Initial state has the
DetailViewController loading syncronous image with some preliminary data.

Next step is definitely to asyncronize the loading of the image. This will
probably require a rewrite of the asyncronous library module. I'll have
to pass in a UIImageView object then populate that instead of adding
my own UIImageView on top of the current UI.