Lots of memory issues resolved. the 11/30 problem
This is the one where I spent 3 hours trying to figure
out the one autorelease pool issue. This one elusive problem
has been mastered. I had to look through all of the release
messages to see which was the culprit. In this case, I was
attempting to release one of the arguments passed into the function.
Line 279 of DishTableViewController.m or nearby was being released.

That problem was rooted in November. Suck.