Updating datastore location as we move
I've set up a background task to handle the movement of the device.
When the user moves, the background task goes through all of the
elements in the core data store and updates the location. So far
no real performance hits. We will need to fine tune how often
this happens though.

https://projects.topdish.com/redmine/issues/91