Fixed up the login flow a little bit
Turns out I was registering for an NSNotification
with self.delegate as the delegate, which is not really
what i wanted. So I was skipping over the LoginModalView from
getting any of the notificiations. I'm not sure when this happened,
but it's so much better this way.

Fixed up the FeedbackViewController to actually send feedback. So that's good.

Quite a few logging reductions as well.

Also, I've added TopDish as a signing team member. That gets us pretty
close to submitting. So that is good news also.

I did have to change com.topdish.TopDish to com.topdish to meet with
the new app id on the provisioning portal.

Lots of changes, very close to submitting.