Hopefully catching all of the networking issues and sending as feedback.
WHen we get some HTML or something wrong, we should be sending
the results to the server as a feedback item. I'm printing the
stacktrace of where it happens along with the response status codes

Also setting the framework for updating the map data from within the
map itself. I'll have a bit of refactoring to do before this is quite ready