Cut down the number of network requests by at least 2/3rds
Removed a lot of unnecessary network calls. This was a long needed task.
Removed the NSNotification stuff. This was dirty an unnecessary.
Changed the accuracy threshold to a bit farther away. It's now nearest hundred meters instead of nearest 10 meters.
Fixed a bug where we were requesting nearby restaurants, then we saw that we were not getting any dishes (duh) and I was expanding the search causing an unwanted network request.
Stopping the location manager periodically so we don't get more network overload. Currently restart location search after 4 minutes.
Adjusted some logic in showing the images on the DishTableView so we get less jumping around images