Adding image cache just saved me from Core Data
I was originally storing the images in Core Data.
The large images were causing the core data fetches to
take an absurd amount of time because it was loading large
image files into memory from disk. Now the cache stores the images
on the disk in the file system and reads them when they are needed.

The speed boost is remarkable on the table view